Init SDL_TIMER

+++ src/gmenu2x.h
@@ -205,7 +205,7 @@ public:
 	InputManager input;
 	Touchscreen ts;
 
-	long tickSuspend; //, tickPowerOff;
+	Uint32 tickSuspend; //, tickPowerOff;
 
 	//Configuration hashes
 	ConfStrHash confStr, skinConfStr;
